+++ server/src/main/java/moment/comment/dto/response/CommentStatusResponse.java
@@ -1,4 +1,4 @@
-package moment.comment.dto;
+package moment.comment.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
